Revised to implement patch request and fixed errors

+++ api/views.py
@@ -7,11 +7,12 @@ from django.shortcuts import get_object_or_404
 from rest_framework import status
 
 # Creating get and post requests
-@api_view(['GET','PUT','DELETE'])
+@api_view(['GET','PUT','DELETE','PATCH'])
 def get_or_update_delete(request,pk):
     
     if request.method=='PUT':
-        serializer=ProductSerializer(data=request.data)
+        product=get_object_or_404(Product,pk=pk)
+        serializer=ProductSerializer(product,data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data,status=status.HTTP_201_CREATED)
@@ -21,10 +22,19 @@ def get_or_update_delete(request,pk):
         product=get_object_or_404(Product,pk=pk)
         serializer=ProductSerializer(product)
         return Response(serializer.data,status=status.HTTP_200_OK)
-    else:
+    elif request.method=='GET':
         product=get_object_or_404(Product,pk=pk)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+    elif request.method=='PATCH':
+        product=get_object_or_404(Product,pk=pk)
+        serializer=ProductSerializer(product,data=request.data,partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data,status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.errors,status=status.HTTP_404_NOT_FOUND)
+        
 
 @api_view(['GET','POST'])
 def product_lists(request):
